requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
proto-plus==1.24.0.dev1
gspread
pandas
